refactor(shared): extract skill manifest helpers into shared.skills
LLM Review and Sub Agent use the same parent-conversation skill manifest parsing and view_skill registration flow. Share it so future skill-capable agent plugins can reuse the same behavior.
